api.js
using node-bitcoin module to interact with the rpc server. It abstracts the calls and makes it much easier than the way I was doing it with the request module.

package.json

adds that module as a dependency

on the angularJS side I updated all the api calls to properly receive the json response.